move choice into utils
