+++ flaml/model.py
@@ -31,7 +31,6 @@ from .data import (
     TOKENCLASSIFICATION,
     SUMMARIZATION,
     NLG_TASKS,
-    MULTICHOICECLASSIFICATION,
 )
 
 try:
